+++ app/Http/Resources/KitResource.php
@@ -5,7 +5,7 @@ namespace App\Http\Resources;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class Kit extends JsonResource
+class KitResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
@@ -19,6 +19,7 @@ class Kit extends JsonResource
             'name' => $this->name,
             'description' => $this->description,
             'image' => $this->image,
+            'created_at' => $this->created_at,
         ];
     }
 }